Extract a few functions to separate mod
